requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.0
python-dotenv==1.0.1

# AI/LLM
openai==1.35.3
anthropic==0.28.0
langchain==0.2.5
langchain-community==0.2.5
langchain-openai==0.1.8

# NLP
transformers==4.41.2
sentence-transformers==3.0.1

# Vector DB
chromadb==0.5.3
faiss-cpu==1.9.0.post1

# Document Processing
pypdf==4.2.0
pdfplumber==0.11.0
pytesseract==0.3.10
pillow==10.4.0
python-multipart==0.0.9

# Data
pandas==2.2.2
numpy==1.26.4
tiktoken==0.7.0

# Security
bcrypt==4.1.3
cryptography==42.0.8

# PDF
reportlab==4.2.0

# API
httpx==0.27.0
